Replaced DISTINCT with GROUP BY.
This seems to have _really_ sped up the query, but once all of the data is wrapped up in all the table markup, there's still a pretty significant amount of data being passed to the browser. This eventually needs to be json-ified so we're passing less data and the browser won't have thousands of table cells to render all at once.